Updates
Add acknowledgments section. Remove modification to [dcl.spec.auto]/13
(as various folks have felt is is problematic, and it is not necessary
to the main objectives of the proposal), but explore it in the
discussion section. Stress somewhat more heavily utility of the main
change (i.e.  inferred return type) that it is useful for all types, not
just anonymous structs. Add section exploring use of anonymous structs
in function pointers. Add future directions section.